+++ python/jaeger/actor/commands/positioner.py
@@ -176,7 +176,7 @@ async def goto(
     except JaegerError as err:
         return command.fail(error=f"Goto command failed: {err}")
 
-    command.finish(text="Position reached")
+    command.finish(text="Destination reached")
 
 
 @jaeger_parser.command()
@@ -363,10 +363,6 @@ async def stop(command, fps):
 @jaeger_parser.command()
 async def unlock(command, fps):
     """Unlocks the FPS."""
-
-    if not fps.locked:
-        command.info(locked=False)
-        return command.finish(text="FPS is not locked")
 
     result = await fps.unlock()